+++ lib/gvc/gvjobs.c
@@ -101,8 +101,7 @@ GVJ_t *gvjobs_next(GVC_t * gvc)
     return (gvc->job = job);
 }
 
-void gv_argvlist_set_item(gv_argvlist_t *list, int index, char *item)
-{
+void gv_argvlist_set_item(gv_argvlist_t *list, size_t index, char *item) {
     if (index >= list->alloc) {
 	list->argv = gv_recalloc(list->argv, list->alloc, index + 10, sizeof(char *));
 	list->alloc = index + 10;
